removed old files, restructured data, refactored about info cards
