--- src/components/users/admin/AdminDashboard.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const AdminDashboard = () => <div>I am the dashboard for an admin!</div>;
-
-export default AdminDashboard;
