--- src/components/users/candidate/CandidateDashboard.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const CandidateDashboard = () => <div>I am the dashboard for a candidate!</div>;
-
-export default CandidateDashboard;
