--- src/components/users/candidate/Profile.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Profile = () => <div>I am the profile for a candidate!</div>;
-
-export default Profile;
