--- src/components/users/instructor/InstructorDashboard.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const InstructorDashboard = () => <div>I am the dashboard for an Instructor!</div>;
-
-export default InstructorDashboard;
