+++ src/data/INFO_CARDS.js
@@ -1,0 +1,23 @@
+const cards = [
+  {
+    img_src: '/images/dndWhite.png',
+    img_alt: 'People',
+    description:
+      'We teach you everything you need to know to start playing Dungeons and Dragons with your friends!',
+  },
+  {
+    img_src: '/images/magicScrollWhite.png',
+    img_alt: 'Person questioning something',
+    description:
+      'Follow our hand-crafted courses to start learning, and start your journey to become the ultimate D&D master!',
+  },
+  {
+    img_src:
+      'https://media.discordapp.net/attachments/540283031001235477/560868650899341351/WhiteJesus.png?width=450&height=450',
+    img_alt: 'Radiating Person',
+    description:
+      'Our goal is to spread the love for D&D by providing an easy-to-learn interface that educates anyone how to play!',
+  },
+];
+
+export default cards;
